+++ src/Styles/home.css
@@ -1,6 +1,6 @@
 @import url('./global.css');
 
 body {
-  background-color: forestgreen;
+  background-color: #262424;
 }
