+++ src/Styles/login.css
@@ -1,7 +1,7 @@
 @import url('./global.css');
 
 body {
-  background-color: red;
+  background-color: #262424;
 }
 
 .form-inputs p {
